clean up sensor icon type config
